+++ apps/content_script/block_target_factory.ts
@@ -13,6 +13,7 @@ const BlockTargetFactory = {
         const blockedSites: BlockedSites = await BlockedSitesRepository.load();
         const bannedWords: IBannedWord[] = await BannedWordRepository.load();
         const idnOption = await OptionRepository.getAutoBlockIDNOption();
+        const defaultBlockType: string = await OptionRepository.defaultBlockType();
         Logger.debug("autoBlockIDNOption:", idnOption);
 
         document.querySelectorAll(".g").forEach((g1: Element) => {
@@ -24,13 +25,13 @@ const BlockTargetFactory = {
 
             const blockState: BlockState = new BlockState(g, blockedSites, bannedWords, idnOption);
 
-            if (blockState.state === "hard") {
+            if (blockState.getState() === "hard") {
                 g.deleteElement();
                 return;
             }
 
             const id = `block${++count}`;
-            const mediator = new BlockMediator(g, blockState, id);
+            const mediator = new BlockMediator(g, blockState, id, defaultBlockType);
         });
 
         document.querySelectorAll("g-inner-card").forEach((g1) => {
@@ -42,13 +43,13 @@ const BlockTargetFactory = {
 
             const blockState: BlockState = new BlockState(g, blockedSites, bannedWords, idnOption);
 
-            if (blockState.state === "hard") {
+            if (blockState.getState() === "hard") {
                 g.deleteElement();
                 return;
             }
 
             const id = `block${++count}`;
-            const mediator = new BlockMediator(g, blockState, id);
+            const mediator = new BlockMediator(g, blockState, id, defaultBlockType);
             mediator.setWrappable("205px");
         });